requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
paddlepaddle==2.6.2
paddleocr==2.7.3
pytesseract==0.3.10
Pillow>=9.0.0
opencv-python<=4.6.0.66
numpy>=1.21.0,<2.0.0
pydantic==2.5.3
python-dotenv==1.0.0
